// types/canvas.ts
export type ActiveElement = {
  name: string;
  value: string;
  icon: string;
} | null;
// ...
export type ShapesMenuProps = {
  item: {
    name: string;
    icon: string;
    value: Array<ActiveElement>;
  };
  activeElement: any;
  handleActiveElement: any;
  // handleImageUpload: any;
  imageInputRef: any;
};